API Reference

CurveFit.CurveFitProblem

CurveFit.CurveFitSolution

CurveFit.ExpSumFitAlgorithm

CurveFit.LinearCurveFitAlgorithm

CurveFit.ModifiedKingCurveFitAlgorithm

CurveFit.PolynomialFitAlgorithm

CurveFit.RationalPolynomialFitAlgorithm

CommonSolve.init

CommonSolve.solve

CommonSolve.solve!

CurveFit.ExpCurveFitAlgorithm

CurveFit.KingCurveFitAlgorithm

CurveFit.LogCurveFitAlgorithm

CurveFit.NonlinearCurveFitProblem

CurveFit.PowerCurveFitAlgorithm

Problem Types

CurveFit.CurveFitProblem — Type

CurveFitProblem(x, y; nlfunc=nothing, u0=nothing)

Represents a curve fitting problem where x and y are the data points to fit.

Certain algorithms may require an initial guess u0 for the coefficients to fit. See specific solver documentation for more details.

See also NonlinearCurveFitProblem.

CurveFit.NonlinearCurveFitProblem — Function

NonlinearCurveFitProblem(f, u0, x, y)

Nonlinear curve fitting problem where f is a nonlinear function to fit, u0 is the initial guess for the coefficients, and x and y are the data points to fit. The following optimization problem is solved:

\[\argmin_u ~ \left\| f(u, x) - y \right\|_2\]

If y is nothing, then it is treated as a zero vector. f is a generic Julia function or ideally a NonlinearFunction from SciMLBase.jl.

Algorithms

CurveFit.LinearCurveFitAlgorithm — Type

LinearCurveFitAlgorithm(;
    xfun = identity, yfun = identity, yfun_inverse = inverse(yfun)
)

Represents a linear curve fitting algorithm where x and y are the data points to fit. We want to solve for a and b such that:

\[f_y(y) = a f_x(x) + b\]

where $f_x$ corresponds to xfun and $f_y$ corresponds to yfun. Note that this is a general problem specification of a curve fitting problem which can be converted to a linear fit in a specific function space by choosing appropriate xfun and yfun. The yfun_inverse is used to convert the fitted values back to the original space (can be specified by defining InverseFunctions.inverse).

CurveFit.PolynomialFitAlgorithm — Type

PolynomialFitAlgorithm(degree::Int)
PolynomialFitAlgorithm(;
    degree::Int,
    linsolve_algorithm::Union{Nothing, AbstractLinearAlgorithm} = nothing
)

Represents a polynomial fitting algorithm of degree degree. Only applicable to LinearCurveFitAlgorithms.

Tip

For ill-conditioned problems, it is recommended to use linear solvers like QRFactorization. Alternatively, pass in assumptions = OperatorAssumptions(false; condition = OperatorsCondition.<condition>) to solve/init.

CurveFit.LogCurveFitAlgorithm — Function

LogCurveFitAlgorithm()

Represents a log curve fitting algorithm where x and y are the data points to fit. We want to solve for a and b such that:

\[y = a \log(x) + b\]

CurveFit.PowerCurveFitAlgorithm — Function

PowerCurveFitAlgorithm()

Represents a power curve fitting algorithm where x and y are the data points to fit. We want to solve for a and b such that:

\[y = b x^a\]

This is equivalent to a linear fit in log-log space, i.e.,

\[\log(y) = a \log(x) + \log(b)\]

CurveFit.ExpCurveFitAlgorithm — Function

ExpCurveFitAlgorithm()

Represents an exponential curve fitting algorithm where x and y are the data points to fit. We want to solve for a and b such that:

\[y = b \exp(a x)\]

This is equivalent to a linear fit in log-linear space, i.e.,

\[\log(y) = a x + \log(b)\]

CurveFit.RationalPolynomialFitAlgorithm — Type

RationalPolynomialFitAlgorithm(num_degree::Int, den_degree::Int)
RationalPolynomialFitAlgorithm(;
    num_degree::Int, den_degree::Int, alg = nothing
)

Represents a rational polynomial fitting algorithm with numerator degree num_degree and denominator degree den_degree. The internal polynomial fitting algorithm is determined by the alg keyword argument. If alg is nothing or a AbstractNonlinearAlgorithm (like solvers from NonlinearSolve.jl), it will use a nonlinear curve fitting approach. If alg is a AbstractLinearAlgorithm, it will use linear least squares fitting.

Linear Rational Polynomial Fitting

In this case the following curve fit is done:

\[y = \frac{p(x)}{q(x)}\]

where p(x) is a polynomial of degree num_degree and q(x) is a polynomial of degree den_degree. The linear case is solved by doing a least squares fit on:

\[y q(x) = p(x)\]

where the zero order term of q(x) is assumed to be 1.

Nonlinear Rational Polynomial Fitting

If an u0 is not provided to the problem, then we will use linear least squares for an initial guess.

CurveFit.KingCurveFitAlgorithm — Function

KingCurveFitAlgorithm()

Represents a king curve fitting problem where x and y are the data points to fit. We want to solve for a and b according to original King's law (1910) that represents the relationship between voltage (E) and velocity (U) in a hotwire anemometer:

\[E^2 = A + B U^{1/2}\]

or

\[x^2 = A + B y^{1/2}\]

CurveFit.ModifiedKingCurveFitAlgorithm — Type

ModifiedKingCurveFitAlgorithm(alg::Union{Nothing, AbstractNonlinearAlgorithm} = nothing)

Similar to KingCurveFitAlgorithm, but uses the modified King's law:

\[E^2 = A + B U^n\]

where n is also a parameter.

CurveFit.ExpSumFitAlgorithm — Type

ExpSumFitAlgorithm(; n::Int, m::Int = 1, withconst::Bool = true)

Fits the sum of n exponentials and a constant.

\[y = k + p_1 e^{λ_1 t} + p_2 e^{λ_2 t} + ⋯ + p_n e^{λ_n t}\]

If the keyword withconst is set to false, the constant is not fitted but set k=0.

Uses numerical integration with m strips, where the default m=1 uses linear interpolation. m=2 and higher require uniform interval and usually lead to better accuracy.

This algorithm is from Matlab code of Juan Gonzales Burgos.

Solutions

CurveFit.CurveFitSolution — Type

CurveFitSolution(alg, coeffs, prob)

Represents the solution to a curve fitting problem. This is a callable struct and can be used to evaluate the solution at a point. Exact evaluation mechanism depends on the algorithm used to solve the problem.

Common Interface

CommonSolve.solve — Function

CommonSolve.solve(args...; kwargs...)

Solves an equation or other mathematical problem using the algorithm specified in the arguments. Generally, the interface is:

CommonSolve.solve(prob::ProblemType,alg::SolverType; kwargs...)::SolutionType

where the keyword arguments are uniform across all choices of algorithms.

By default, solve defaults to using solve! on the iterator form, i.e.:

solve(args...; kwargs...) = solve!(init(args...; kwargs...))

CommonSolve.solve! — Function

CommonSolve.solve!(iter)

Solves an equation or other mathematical problem using the algorithm specified in the arguments. Generally, the interface is:

iter = CommonSolve.init(prob::ProblemType,alg::SolverType; kwargs...)::IterType
CommonSolve.solve!(iter)::SolutionType

where the keyword arguments are uniform across all choices of algorithms. The iter type will be different for the different problem types.

CommonSolve.init — Function

iter = CommonSolve.init(args...; kwargs...)

Solves an equation or other mathematical problem using the algorithm specified in the arguments. Generally, the interface is:

iter = CommonSolve.init(prob::ProblemType,alg::SolverType; kwargs...)::IterType
CommonSolve.solve!(iter)::SolutionType

where the keyword arguments are uniform across all choices of algorithms. The iter type will be different for the different problem types.